% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epub-to-text.R
\name{epub_to_text}
\alias{epub_to_text}
\title{Convert an epub file into a data frame of plaintext chapters}
\usage{
epub_to_text(path)
}
\arguments{
\item{path}{path to epub}
}
\description{
Convert an epub file into a data frame of plaintext chapters
}
\examples{
\dontrun{
epub_to_text("~/Data/R Packages.epub")
}
}
