% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubcrawl-package.R
\docType{package}
\name{pubcrawl}
\alias{pubcrawl}
\alias{pubcrawl-package}
\title{Convert 'epub' Files to Text}
\description{
The 'epub' file format is really just a structured 'ZIP' archive with
metadata, graphics and (usually) 'HTML' text. Tools are provided to turn an 'epub'
file into a tidy data frame.
}
\author{
Bob Rudis (bob@rud.is)
}
