% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epub-to-text.R
\name{epub_to_text}
\alias{epub_to_text}
\title{Convert an epub file into a data frame of plaintext chapters}
\usage{
epub_to_text(path, encoding = "UTF-8")
}
\arguments{
\item{path}{path to epub}

\item{encoding}{Some content may require special handling of
encodings. The default (`UTF-8`) can be overriden with this parameter.}
}
\description{
Convert an epub file into a data frame of plaintext chapters
}
\examples{
epub_to_text(system.file("extdat", "augustine.epub", package="pubcrawl"))
}
